add beacon count

+++ dot11.h
@@ -9,6 +9,6 @@ typedef struct radiotap_header {
 
 typedef struct beacon {
         u_int8_t        bssid[6];
-        u_int8_t        bc_cnt;
+        int             bc_cnt;
         char            essid[256];
 }BEACON;
